Change format for spec harness
I had to change the IDs to integers - and make sure that I was always
saving them as integers and searching for them with integers.

I also had to change the price to save in dollars instead of cents.

Both changes were to match choices made in the Spec Harness.